+++ src/Theme/ThemeBase.php
@@ -50,6 +50,26 @@ abstract class ThemeBase
         $this->serviceProvider->setViews($this->getPath() . $this->viewsPath . $this->templatesFolder);
     }
 
+    public function install() : void
+    {
+        //
+    }
+
+    public function uninstall() : void
+    {
+        //
+    }
+
+    public function activate() : void
+    {
+        //
+    }
+
+    public function deactivate() : void
+    {
+        //
+    }
+
     public function availableViews() : Collection
     {
         if ($this->availableViews === null) {