+++ tests/ThemeManagerTest.php
@@ -4,6 +4,7 @@ namespace Aldrumo\ThemeManager\Tests;
 
 use Aldrumo\ThemeManager\Exceptions\ActiveThemeNotSetException;
 use Aldrumo\ThemeManager\Exceptions\ThemeNotFoundException;
+use Aldrumo\ThemeManager\Tests\TestClasses\AnotherTheme;
 use Aldrumo\ThemeManager\Tests\TestClasses\AnotherThemeServiceProvider;
 use Aldrumo\ThemeManager\Tests\TestClasses\DefaultTheme;
 use Aldrumo\ThemeManager\Tests\TestClasses\DefaultThemeServiceProvider;
@@ -72,5 +73,75 @@ class ThemeManagerTest extends TestCase
         $themes = app(ThemeManager::class)->availableThemes();
 
         $this->assertEmpty($themes->toArray());
+    }
+
+    /** @test */
+    public function can_switch_theme()
+    {
+        $this->bootThemes();
+        app(ThemeManager::class)->activeTheme('DefaultTheme');
+
+        app(ThemeManager::class)->activateTheme('AnotherTheme', 'DefaultTheme');
+
+        $this->assertInstanceOf(
+            AnotherTheme::class,
+            app(ThemeManager::class)->activeTheme()
+        );
+    }
+
+    /** @test */
+    public function can_activate_theme()
+    {
+        $this->bootThemes();
+        app(ThemeManager::class)->activateTheme('AnotherTheme');
+
+        $this->assertInstanceOf(
+            AnotherTheme::class,
+            app(ThemeManager::class)->activeTheme()
+        );
+    }
+
+    /** @test */
+    public function activate_theme_throws_exception_when_invalid_old_theme_set()
+    {
+        $this->bootThemes();
+        app(ThemeManager::class)->activeTheme('DefaultTheme');
+
+        $this->expectException(ThemeNotFoundException::class);
+        app(ThemeManager::class)->activateTheme('AnotherTheme', 'ThemeDoesNotExist');
+    }
+
+    /** @test */
+    public function activate_theme_throws_exception_when_invalid_new_theme_set()
+    {
+        $this->bootThemes();
+        app(ThemeManager::class)->activeTheme('DefaultTheme');
+
+        $this->expectException(ThemeNotFoundException::class);
+        app(ThemeManager::class)->activateTheme('ThemeDoesNotExist', 'DefaultTheme');
+    }
+
+    /** @test */
+    public function can_install_theme()
+    {
+        $this->bootThemes();
+        $theme = app(ThemeManager::class)->installTheme('AnotherTheme');
+
+        $this->assertInstanceOf(
+            AnotherTheme::class,
+            $theme
+        );
+    }
+
+    /** @test */
+    public function can_uninstall_theme()
+    {
+        $this->bootThemes();
+        $theme = app(ThemeManager::class)->uninstallTheme('AnotherTheme');
+
+        $this->assertInstanceOf(
+            AnotherTheme::class,
+            $theme
+        );
     }
 }